Add survey inheritance and update categories in module manifest
This commit includes the creation of new inherited views for survey and survey questions in XML. Additionally, it updates the module category in __manifest__.py file from 'Uncategorized' to 'Marketing' and adds the newly created views to the data list. The 'survey.survey' and 'survey.question' models are also augmented with new fields.